Tabla PedidoItem e integracion con Pedidos

+++ back/API/DTOs/Pedido/PedidoByUserIdResponseDto.cs
@@ -7,8 +7,8 @@
         public string EmailCliente { get; set; } = string.Empty;
         public string TelefonoCliente { get; set; } = string.Empty;
         public string Direccion { get; set; } = string.Empty;
-        public double Subtotal { get; set; }
-        public double Total { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal Total { get; set; }
         public DateTime PedidoFecha { get; set; } = DateTime.UtcNow;
         public DateTime FechaEntrega { get; set; } = DateTime.UtcNow.AddDays(7);
